[chore] implement GPA and GPAX calculation in TermSummaryUsecase
